git_ranger/requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
chardet==3.0.4
cryptography==2.4.2
Deprecated==1.2.4
dict2xml==1.5
Django==2.1.3
django-prometheus==1.0.15
django-reset-migrations==0.3.1
djangorestframework==3.9.0
idna==2.7
json2xml==2.2.1
prometheus-client==0.4.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyAesCrypt==0.4.2
pycparser==2.19
PyGithub==1.43.3
PyJWT==1.6.4
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
wrapt==1.10.11
xmltodict==0.11.0